CellTrackingBase: moved file to demo/advanced
